Yakim 478

+++ project/Starter/src/main/java/practics/Leambda/Incrementer.java
@@ -1,5 +1,5 @@
 package practics.Leambda;
 
 public interface Incrementer {
-    int increment(int x);
+    double increment(int x);
 }
